fixing flataction discriminator to convert to float beforehand, adding necessary forward calls in expert, adding Fire to video creator from model, and trying full run of options gail with new discrimination model

+++ scratch/etienne/intersimple/gail/discriminator.py
@@ -75,7 +75,7 @@ class CnnDiscriminatorFlatAction(torch.nn.Module):
         return sa
 
     def forward(self, state, action):
-        s = self.cnn(state)
+        s = self.cnn(state.float())
         sa = self._concatenate(s, action)
         assert sa.ndim == 2
         return self.decoder(sa).squeeze(1)
